+++ packages/devon/src/templates/.devconfig.ts
@@ -1,4 +1,4 @@
-export const serviceFile = `import type { ServiceConfig } from '@casthub/devon';
+export const serviceFile = `import type { ServiceConfig } from '@tnotifier/devon';
 
 export const config: ServiceConfig = {
     // Put your env, docker-compose etc. config here!